Fix tab label colors for Origin and Sodish
The background of the tab_control of Origin and Sodish is not derived
from the editor control's background but from the main window.

This commit applies the same rule to the tab_label, because otherwise
an editor with light background can make the label barely visible if
the main window is otherwise dark.

Note:
  Looks like the `file_light` attribute of the `tab_control` is not
  calulated from its actual background color but from the active
  view's one.