Base: Split sidebar vcs folder and file label rules
The label colors of folders and files are now defined by separate sets
of rules to have better control about how folders and files are tinted
in different situations.

Tinting folders which contain ignored or untracked files is for instance
not optimal as it looks like the whole folder was untracked/ignored if
all other files are committed.