All: Improve highlighted tab labels
ST 4095+ synchronizes state attributes between a view and its tab,
which causes tab labels of the active and highlighted tabs changing
color while moving the cursor within the view in order to select text
etc. That's quite annoying.

In order to reduce that effect and to emphasize the one and only
highlighted view of a window, the "highlighted" tab is assigned its own
label color (which is the same as the hover color by default) and the
label is rendered bold.